+++ plutarch-ledger-api/src/Plutarch/LedgerApi/V1.hs
@@ -105,6 +105,8 @@ data PTxOut (s :: S) = PTxOut
       PEq
     , -- | @since 3.1.1
       PShow
+    , -- | @since 3.3.1
+      PTryFrom PData
     )
   deriving
     ( -- | @since 3.3.0
@@ -117,6 +119,9 @@ deriving via
   DeriveDataPLiftable PTxOut Plutus.TxOut
   instance
     PLiftable PTxOut
+
+-- | @since 3.3.1
+instance PTryFrom PData (PAsData PTxOut)
 
 -- | @since 3.1.1
 data PTxInInfo (s :: S) = PTxInInfo
@@ -136,6 +141,8 @@ data PTxInInfo (s :: S) = PTxInInfo
       PEq
     , -- | @since 3.1.1
       PShow
+    , -- | @since 3.3.1
+      PTryFrom PData
     )
   deriving
     ( -- | @since 3.3.0
@@ -148,6 +155,9 @@ deriving via
   DeriveDataPLiftable PTxInInfo Plutus.TxInInfo
   instance
     PLiftable PTxInInfo
+
+-- | @since 3.3.1
+instance PTryFrom PData (PAsData PTxInInfo)
 
 -- | @since 3.3.0
 data PTxInfo (s :: S) = PTxInfo
@@ -175,6 +185,8 @@ data PTxInfo (s :: S) = PTxInfo
       PEq
     , -- | @since 3.1.1
       PShow
+    , -- | @since 3.3.1
+      PTryFrom PData
     )
   deriving
     ( -- | @since 3.3.0
@@ -187,6 +199,9 @@ deriving via
   DeriveDataPLiftable PTxInfo Plutus.TxInfo
   instance
     PLiftable PTxInfo
+
+-- | @since 3.3.1
+instance PTryFrom PData (PAsData PTxInfo)
 
 -- | @since 3.1.1
 data PScriptContext (s :: S) = PScriptContext
@@ -206,6 +221,8 @@ data PScriptContext (s :: S) = PScriptContext
       PEq
     , -- | @since 3.1.1
       PShow
+    , -- | @since 3.3.1
+      PTryFrom PData
     )
   deriving
     ( -- | @since 3.3.0
@@ -218,3 +235,6 @@ deriving via
   DeriveDataPLiftable PScriptContext Plutus.ScriptContext
   instance
     PLiftable PScriptContext
+
+-- | @since 3.3.1
+instance PTryFrom PData (PAsData PScriptContext)