+++ plutarch-ledger-api/src/Plutarch/LedgerApi/V1/Contexts.hs
@@ -33,6 +33,8 @@ data PScriptPurpose (s :: S)
       PEq
     , -- | @since 3.1.1
       PShow
+    , -- | @since 3.3.1
+      PTryFrom PData
     )
   deriving
     ( -- | @since 3.3.0
@@ -45,3 +47,6 @@ deriving via
   DeriveDataPLiftable PScriptPurpose Plutus.ScriptPurpose
   instance
     PLiftable PScriptPurpose
+
+-- | @since 3.3.1
+instance PTryFrom PData (PAsData PScriptPurpose)