+++ plutarch-ledger-api/src/Plutarch/LedgerApi/V3/Contexts.hs
@@ -74,6 +74,8 @@ newtype PColdCommitteeCredential (s :: S) = PColdCommitteeCredential (Term s PCr
       PEq
     , -- | @since 3.1.0
       PShow
+    , -- | @since 3.3.1
+      PTryFrom PData
     )
   deriving
     ( -- | @since 3.3.0
@@ -86,6 +88,9 @@ deriving via
   DeriveDataPLiftable PColdCommitteeCredential Plutus.ColdCommitteeCredential
   instance
     PLiftable PColdCommitteeCredential
+
+-- | @since 3.3.1
+instance PTryFrom PData (PAsData PColdCommitteeCredential)
 
 -- | @since 3.1.0
 newtype PHotCommitteeCredential (s :: S) = PHotCommitteeCredential (Term s PCredential)
@@ -102,6 +107,8 @@ newtype PHotCommitteeCredential (s :: S) = PHotCommitteeCredential (Term s PCred
       PEq
     , -- | @since 3.1.0
       PShow
+    , -- | @since 3.3.1
+      PTryFrom PData
     )
   deriving
     ( -- | @since 3.3.0
@@ -114,6 +121,9 @@ deriving via
   DeriveDataPLiftable PHotCommitteeCredential Plutus.HotCommitteeCredential
   instance
     PLiftable PHotCommitteeCredential
+
+-- | @since 3.3.1
+instance PTryFrom PData (PAsData PHotCommitteeCredential)
 
 -- | @since 3.1.0
 newtype PDRepCredential (s :: S) = PDRepCredential (Term s PCredential)
@@ -130,6 +140,8 @@ newtype PDRepCredential (s :: S) = PDRepCredential (Term s PCredential)
       PEq
     , -- | @since 3.1.0
       PShow
+    , -- | @since 3.3.1
+      PTryFrom PData
     )
   deriving
     ( -- | @since 3.3.0
@@ -142,6 +154,9 @@ deriving via
   DeriveDataPLiftable PDRepCredential Plutus.DRepCredential
   instance
     PLiftable PDRepCredential
+
+-- | @since 3.3.1
+instance PTryFrom PData (PAsData PDRepCredential)
 
 -- | @since 3.1.0
 data PDRep (s :: S)
@@ -161,6 +176,8 @@ data PDRep (s :: S)
       PEq
     , -- | @since 3.1.0
       PShow
+    , -- | @since 3.3.1
+      PTryFrom PData
     )
   deriving
     ( -- | @since 3.3.0
@@ -173,6 +190,9 @@ deriving via
   DeriveDataPLiftable PDRep Plutus.DRep
   instance
     PLiftable PDRep
+
+-- | @since 3.3.1
+instance PTryFrom PData (PAsData PDrep)
 
 -- | @since 3.1.0
 data PDelegatee (s :: S)
@@ -192,6 +212,8 @@ data PDelegatee (s :: S)
       PEq
     , -- | @since 3.1.0
       PShow
+    , -- | @since 3.3.1
+      PTryFrom PData
     )
   deriving
     ( -- | @since 3.3.0
@@ -204,6 +226,9 @@ deriving via
   DeriveDataPLiftable PDelegatee Plutus.Delegatee
   instance
     PLiftable PDelegatee
+
+-- | @since 3.3.1
+instance PTryFrom PData (PAsData PDelegatee)
 
 -- | @since 3.1.0
 data PTxCert (s :: S)
@@ -231,6 +256,8 @@ data PTxCert (s :: S)
       PEq
     , -- | @since 3.1.0
       PShow
+    , -- | @since 3.3.1
+      PTryFrom PData
     )
   deriving
     ( -- | @since 3.3.0
@@ -243,6 +270,9 @@ deriving via
   DeriveDataPLiftable PTxCert Plutus.TxCert
   instance
     PLiftable PTxCert
+
+-- | @since 3.3.1
+instance PTryFrom PData (PAsData PTxCert)
 
 -- | @since 3.1.0
 data PVoter (s :: S)
@@ -262,6 +292,8 @@ data PVoter (s :: S)
       PEq
     , -- | @since 3.1.0
       PShow
+    , -- | @since 3.3.1
+      PTryFrom PData
     )
   deriving
     ( -- | @since 3.3.0
@@ -274,6 +306,9 @@ deriving via
   DeriveDataPLiftable PVoter Plutus.Voter
   instance
     PLiftable PVoter
+
+-- | @since 3.3.1
+instance PTryFrom PData (PAsData PVoter)
 
 -- | @since 3.1.0
 data PVote (s :: S)
@@ -293,6 +328,8 @@ data PVote (s :: S)
       PEq
     , -- | @since 3.1.0
       PShow
+    , -- | @since 3.3.1
+      PTryFrom PData
     )
   deriving
     ( -- | @since 3.3.0
@@ -305,6 +342,9 @@ deriving via
   DeriveDataPLiftable PVote Plutus.Vote
   instance
     PLiftable PVote
+
+-- | @since 3.3.1
+instance PTryFrom PData (PAsData PVote)
 
 -- | @since 3.1.0
 data PGovernanceActionId (s :: S)
@@ -322,6 +362,8 @@ data PGovernanceActionId (s :: S)
       PEq
     , -- | @since 3.1.0
       PShow
+    , -- | @since 3.3.1
+      PTryFrom PData
     )
   deriving
     ( -- | @since 3.3.0
@@ -334,6 +376,9 @@ deriving via
   DeriveDataPLiftable PGovernanceActionId Plutus.GovernanceActionId
   instance
     PLiftable PGovernanceActionId
+
+-- | @since 3.3.1
+instance PTryFrom PData (PAsData PGovernanceActionId)
 
 -- TODO: Investigate what guarantees this provides on the Map, if any
 
@@ -355,6 +400,8 @@ data PCommittee (s :: S) = PCommittee
       PEq
     , -- | @since 3.1.0
       PShow
+    , -- | @since 3.3.1
+      PTryFrom PData
     )
   deriving
     ( -- | @since 3.3.0
@@ -367,6 +414,9 @@ deriving via
   DeriveDataPLiftable PCommittee Plutus.Committee
   instance
     PLiftable PCommittee
+
+-- | @since 3.3.1
+instance PTryFrom PData (PAsData PCommittee)
 
 {- | A constitution, omitting the optional anchor.
 
@@ -386,6 +436,8 @@ newtype PConstitution (s :: S) = PConstitution (Term s (PMaybeData PScriptHash))
       PEq
     , -- | @since 3.1.0
       PShow
+    , -- | @since 3.3.1
+      PTryFrom PData
     )
   deriving
     ( -- | @since 3.3.0
@@ -398,6 +450,9 @@ deriving via
   DeriveDataPLiftable PConstitution Plutus.Constitution
   instance
     PLiftable PConstitution
+
+-- | @since 3.3.1
+instance PTryFrom PData (PAsData PConstitution)
 
 -- | @since 3.1.0
 data PProtocolVersion (s :: S) = PProtocolVersion
@@ -417,6 +472,8 @@ data PProtocolVersion (s :: S) = PProtocolVersion
       PEq
     , -- | @since 3.1.0
       PShow
+    , -- | @since 3.3.1
+      PTryFrom PData
     )
   deriving
     ( -- | @since 3.3.0
@@ -429,6 +486,9 @@ deriving via
   DeriveDataPLiftable PProtocolVersion Plutus.ProtocolVersion
   instance
     PLiftable PProtocolVersion
+
+-- | @since 3.3.1
+instance PTryFrom PData (PAsData PProtocolVersion)
 
 -- | @since 3.1.0
 newtype PChangedParameters (s :: S)
@@ -446,6 +506,8 @@ newtype PChangedParameters (s :: S)
       PEq
     , -- | @since 3.1.0
       PShow
+    , -- | @since 3.3.1
+      PTryFrom PData
     )
   deriving
     ( -- | @since 3.3.0
@@ -458,6 +520,9 @@ deriving via
   DeriveDataPLiftable PChangedParameters Plutus.ChangedParameters
   instance
     PLiftable PChangedParameters
+
+-- | @since 3.3.1
+instance PTryFrom PData (PAsData PChangedParameters)
 
 -- | @since 3.1.0
 data PGovernanceAction (s :: S)
@@ -487,6 +552,8 @@ data PGovernanceAction (s :: S)
       PEq
     , -- | @since 3.1.0
       PShow
+    , -- | @since 3.3.1
+      PTryFrom PData
     )
   deriving
     ( -- | @since 3.3.0
@@ -499,6 +566,9 @@ deriving via
   DeriveDataPLiftable PGovernanceAction Plutus.GovernanceAction
   instance
     PLiftable PGovernanceAction
+
+-- | @since 3.3.1
+instance PTryFrom PData (PAsData PGovernanceAction)
 
 -- | @since 3.1.0
 data PProposalProcedure (s :: S) = PProposalProcedure
@@ -519,6 +589,8 @@ data PProposalProcedure (s :: S) = PProposalProcedure
       PEq
     , -- | @since 3.1.0
       PShow
+    , -- | @since 3.3.1
+      PTryFrom PData
     )
   deriving
     ( -- | @since 3.3.0
@@ -531,6 +603,9 @@ deriving via
   DeriveDataPLiftable PProposalProcedure Plutus.ProposalProcedure
   instance
     PLiftable PProposalProcedure
+
+-- | @since 3.3.1
+instance PTryFrom PData (PAsData PProposalProcedure)
 
 -- | @since 2.0.0
 data PScriptPurpose (s :: S)
@@ -556,6 +631,8 @@ data PScriptPurpose (s :: S)
       PEq
     , -- | @since 2.0.0
       PShow
+    , -- | @since 3.3.1
+      PTryFrom PData
     )
   deriving
     ( -- | @since 3.3.0
@@ -568,6 +645,9 @@ deriving via
   DeriveDataPLiftable PScriptPurpose Plutus.ScriptPurpose
   instance
     PLiftable PScriptPurpose
+
+-- | @since 3.3.1
+instance PTryFrom PData (PAsData PScriptPurpose)
 
 -- | @since 3.1.0
 data PScriptInfo (s :: S)
@@ -590,6 +670,8 @@ data PScriptInfo (s :: S)
       PEq
     , -- | @since 3.1.0
       PShow
+    , -- | @since 3.3.1
+      PTryFrom PData
     )
   deriving
     ( -- | @since 3.3.0
@@ -602,6 +684,9 @@ deriving via
   DeriveDataPLiftable PScriptInfo Plutus.ScriptInfo
   instance
     PLiftable PScriptInfo
+
+-- | @since 3.3.1
+instance PTryFrom PData (PAsData PScriptInfo)
 
 {- | An input of the pending transaction.
 
@@ -624,6 +709,8 @@ data PTxInInfo (s :: S) = PTxInInfo
       PEq
     , -- | @since 2.0.0
       PShow
+    , -- | @since 3.3.1
+      PTryFrom PData
     )
   deriving
     ( -- | @since 3.3.0
@@ -636,6 +723,9 @@ deriving via
   DeriveDataPLiftable PTxInInfo Plutus.TxInInfo
   instance
     PLiftable PTxInInfo
+
+-- | @since 3.3.1
+instance PTryFrom PData (PAsData PTxInInfo)
 
 -- A pending transaction. This is the view as seen by a validator script.
 --
@@ -671,6 +761,8 @@ data PTxInfo (s :: S) = PTxInfo
       PEq
     , -- | @since 2.0.0
       PShow
+    , -- | @since 3.3.1
+      PTryFrom PData
     )
   deriving
     ( -- | @since 3.3.0
@@ -683,6 +775,9 @@ deriving via
   DeriveDataPLiftable PTxInfo Plutus.TxInfo
   instance
     PLiftable PTxInfo
+
+-- | @since 3.3.1
+instance PTryFrom PData (PAsData PTxInfo)
 
 -- | @since 3.1.0
 data PScriptContext (s :: S) = PScriptContext
@@ -703,6 +798,8 @@ data PScriptContext (s :: S) = PScriptContext
       PEq
     , -- | @since 2.0.0
       PShow
+    , -- | @since 3.3.1
+      PTryFrom PData
     )
   deriving
     ( -- | @since 3.3.0
@@ -715,6 +812,9 @@ deriving via
   DeriveDataPLiftable PScriptContext Plutus.ScriptContext
   instance
     PLiftable PScriptContext
+
+-- | @since 3.3.1
+instance PTryFrom PData (PAsData PScriptContext)
 
 {- | Find the datum corresponding to a datum hash, if there is one.